Move dialect exported logic to client module. Remove amf-tools module
