+++ amf-client/jvm/src/main/scala/amf/tools/canonical/CanonicalWebAPITransformer.scala
